pyproject.toml
[project]
name = "article-ai"
version = "0.1.0"
description = "Replication package for article-ai analyses"
requires-python = ">=3.11,<3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "mypy>=1.18.2",
    "openai>=2.1.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.14.1",
    "scons>=4.10.1",
]


[tool.uv]
package = false
